refactor(core): move core related code to core.js.
